[Store edit] + [Store delete] + [cascade relationships] + ![delete products after deleting stores]

+++ src/main/java/marks/webstore/repos/ProductTypeStoreRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 
 public interface ProductTypeStoreRepo extends CrudRepository<ProductTypeStore, Long> {
-    List<ProductTypeStore> findAllByStoreId(Integer storeId);
+    List<ProductTypeStore> findAllByStoreId(Long storeId);
     List<ProductTypeStore> findAllByProductId(Long productId);
     List<ProductTypeStore> findAll();
 }
